Got tests setup and fixed bug.
This should really be two seperate commits: the stuff that got the tests
setup and the bug regarding the parsing of subgroup statements.

In regards to the bug, basically the parser would look at `"*text*"` and think
that the second `*` was the beginning of another italic subgroup within the
first italic subgroup.